# remove duplication of parse logic in UnmarshalJSON