Remove the duplicate parse logic in `UnmarshalJSON`; use the `parse` function instead. Add a benchmark for `UnmarshalJSON` to measure change in performance.

Here's the [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) output comparing the old and new `UnmarshalJSON` code (run on a 2021 14" Macbook Pro, M1 Pro):
```
goos: darwin
goarch: arm64
pkg: github.com/woodsbury/decimal128
                                                                        │   old.txt   │               new.txt               │
                                                                        │   sec/op    │   sec/op     vs base                │
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005-8           49.54n ± 0%   50.53n ± 0%   +2.00% (p=0.000 n=10)
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e30-8        51.50n ± 0%   52.20n ± 0%   +1.37% (p=0.000 n=10)
DecimalUnmarshalJSON/500000000000000000000000000000000000000e30-8         129.4n ± 0%   137.1n ± 0%   +5.95% (p=0.000 n=10)
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-6150-8     65.86n ± 0%   66.25n ± 0%   +0.59% (p=0.000 n=10)
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-999999-8   50.09n ± 0%   52.22n ± 0%   +4.25% (p=0.000 n=10)
DecimalUnmarshalJSON/0e99999999-8                                         14.59n ± 0%   16.45n ± 0%  +12.75% (p=0.000 n=10)
DecimalUnmarshalJSON/0-8                                                  15.09n ± 0%   13.31n ± 3%  -11.83% (p=0.000 n=10)
DecimalUnmarshalJSON/+1234567890-8                                        17.69n ± 0%   19.55n ± 0%  +10.54% (p=0.000 n=10)
DecimalUnmarshalJSON/00123.45600e10-8                                     23.72n ± 0%   25.44n ± 0%   +7.23% (p=0.000 n=10)
DecimalUnmarshalJSON/123.456e-10-8                                        23.14n ± 0%   25.12n ± 0%   +8.54% (p=0.000 n=10)
geomean                                                                   34.41n        35.76n        +3.92%

                                                                        │   old.txt    │               new.txt               │
                                                                        │     B/op     │    B/op     vs base                 │
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e30-8        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/500000000000000000000000000000000000000e30-8         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-6150-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-999999-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0e99999999-8                                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0-8                                                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/+1234567890-8                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/00123.45600e10-8                                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/123.456e-10-8                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                              ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                                        │   old.txt    │               new.txt               │
                                                                        │  allocs/op   │ allocs/op   vs base                 │
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005-8           0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e30-8        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/500000000000000000000000000000000000000e30-8         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-6150-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0.00000000000000000000000000000000000005e-999999-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0e99999999-8                                         0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/0-8                                                  0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/+1234567890-8                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/00123.45600e10-8                                     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
DecimalUnmarshalJSON/123.456e-10-8                                        0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                              ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```