@woodsbury 2023-05-20
I'm uncertain about this change. Reducing the amount of duplicated code is good, but there is a subtle difference in what's considered valid between `Parse` and `UnmarshalJSON` (basically whether a separator character is allowed, along with the infinity and NaN parsing) - I know that if it's called via `json.Unmarshal` it should already have been validated, but I also know there's code out there that effectively calls `UnmarshalJSON` directly.

I have an idea that may allow us to keep the different validation behaviour while still sharing more of the code between the two functions (and maybe `Scan` as well?). I might have a play with it and see if I can get what I'm thinking to work.