@matthewdale 2023-05-21
@woodsbury thanks for the feedback! That's a great point about the different expected inputs using `UnmarshalJSON` directly vs using `json.Unmarshal`. It sounds like adding some test cases for the JSON-specific unmarshal/parse behavior would helpful in any case. I'll try to put together a separate PR for additional test cases. Let me know if you're able to get your idea for consolidating the parse logic working.

Also, agreed that the duplication in `Scan` would be helpful to remove. I've been working on a refactor of `Scan` to use `parse`, which I will submit as soon as it's done. It mostly uses the same approach as this PR, using [fmt.ScanState.Token](https://pkg.go.dev/fmt#ScanState) to extract the next decimal128 token and passing it to `parse`.