@woodsbury 2023-05-22
I've created a branch `feature/parse` with the changes I mentioned before, to combine the parsing logic used by `Parse`/etc and `UnmarshalJSON`.

Essentially, moving the actual number parsing out of `parse()` (and leave the infinity and NaN parsing behind) and into a new function that can be shared.

`Scan` also has some idiosyncrasies in how it parses infinity and NaN, but my hope is that, after those special values have been handled, this new function can also be used with `ScanState.Token` like you mentioned.

I added a handful of extra tests for `UnmarshalJSON` while I was at it, but I'm sure there's a lot more testing that could be done in that area, especially negative test cases. I also haven't tried any benchmarks with my changes yet.