@woodsbury 2023-05-27
I've merged some more changes now to also remove the duplicated parsing logic for `Scan`.