@matthewdale 2023-06-03
@woodsbury your changes to reduce parse logic duplication look great! They seem to remove the need for most of the changes in this PR, so I'll close it. I can contribute the benchmarks and tests separately.